requirements.txt
openai
opencv-python
pillow
torch
torchvision
transformers
pandas
python-docx
fastapi
uvicorn
ultralytics
tk
